Moves configMap functions into their own module
